backend/requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.1
Flask_Caching==2.3.1
flask_cors==5.0.1
fuzzywuzzy==0.18.0
huggingface_hub==0.30.2
langcodes==3.5.0
langdetect==1.0.9
numpy==2.1.3
python-dotenv==1.1.1
Requests==2.32.4
torch==2.7.0
transformers[sentencepiece]==4.51.3
sacremoses==0.1.1